# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Catálogo Publiex Maracaibo
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uploads:
    post:
      operationId: uploadProductImages
      tags: [products]
      summary: Upload product images
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [images]
              properties:
                images:
                  type: array
                  maxItems: 10
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Uploaded images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadImagesResponse"
        "400":
          description: Invalid images
        "401":
          description: Authentication required
  /uploads/sign:
    post:
      operationId: createSignedUpload
      tags: [products]
      summary: Create a signed Storage upload
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadSignInput"
      responses:
        "200":
          description: Signed upload or local development fallback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadSignResponse"
        "400":
          description: Invalid image metadata
        "401":
          description: Authentication required
  /uploads/complete:
    post:
      operationId: completeSignedUpload
      tags: [products]
      summary: Verify a completed Storage upload
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadCompleteInput"
      responses:
        "200":
          description: Verified image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadImagesResponseImagesItem"
        "400":
          description: Invalid or missing image
        "401":
          description: Authentication required
...
components:
  schemas:
...
    UploadImagesResponse:
      type: object
      required: [images]
      properties:
        images:
          type: array
          items:
            $ref: "#/components/schemas/UploadImagesResponseImagesItem"
    UploadImagesResponseImagesItem:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
        url:
          type: string
    UploadSignInput:
      type: object
      required: [kind, name, contentType, size]
      properties:
        kind:
          type: string
          enum: [product, banner]
        name:
          type: string
          maxLength: 255
        contentType:
          type: string
          enum: [image/avif, image/gif, image/jpeg, image/png, image/webp]
        size:
          type: integer
          minimum: 1
          maximum: 5242880
    UploadSignResponse:
      oneOf:
        - type: object
          required: [mode, kind, bucket, storagePath, uploadUrl, publicUrl]
          properties:
            mode:
              type: string
              enum: [signed]
            kind:
              type: string
              enum: [product, banner]
            bucket:
              type: string
            storagePath:
              type: string
            uploadUrl:
              type: string
            publicUrl:
              type: string
        - type: object
          required: [mode]
          properties:
            mode:
              type: string
              enum: [legacy]
    UploadCompleteInput:
      allOf:
        - $ref: "#/components/schemas/UploadSignInput"
        - type: object
          required: [storagePath]
          properties:
            storagePath:
              type: string
              maxLength: 300
...